Improve Value Converters (#227)
* Improve Value Converters
Additional configuration and features in using Value Converters have been implemented and various issue fixes with existing value converter implementation.
- Created reference to a FieldExpression in LiteralExpression, AssignmentExpression, and InsertExpression to have a direct correlation between the source FieldExpression and the using Expression.
- Deprecated FieldExpression from AssemblyContext.  As FieldExpression is now part of the creation of Expressions related to the FieldExpression, using the Push/Pop methods for a FieldExpression during assembly were no longer needed.
- Refactored parameter builders to provide consistency across all creation methods by always returning a ParameterizedExpression.
- Refactored function expressions to "hide" properties via IExpressionProvider implementation.  This ensures no confusion in the API when used (i.e. db.fx.DatePart(year, date1, date1).DatePart is confusing)

Other
- Cosmetic change for assembling SQL statements: removed extraneous indentation while building queries containing offset and limit specifications.

+++ test/HatTrick.DbEx.MsSql.Test/_TypeCode/AddressType.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace DbEx.Data
 {
-    public enum AddressType : int
+    public enum AddressType : long
 	{
 		[Display(Name = "Shipping", Description = "Shipping Address")]
 		Shipping = 0,
